Change #1
When I first opened this file it was empty with the exception that there was a sentence on the first header that said that I had to change it, so I did it. I added my name and a few more headers with different font sizes. The second one with an <h3> specifies the degree I am studying and the third one with <h6> wishes that the reader has a nice time while it is on the website.